[REF] 관습적으로 서비스/impl로 나눠놓은 서비스 레이어 통합
